app/requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3
Werkzeug>=2.0.0